\name{fMM-package}
\alias{fMM-package}
\alias{fMM}
\docType{package}
\title{
\packageTitle{fMM}
}
\description{
\packageDescription{fMM}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{fMM}
\packageIndices{fMM}
Efficient matix mutliplication for big datasets.
}
\author{
\packageAuthor{fMM}
Maintainer: \packageMaintainer{fMM}
}
\references{
Find reference.
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\examples{
~~ simple examples of the most important functions ~~
}
