\name{fMM-package}
\alias{fMM-package}
\alias{fMM}
\docType{package}
\title{
\packageTitle{fMM}
}
\description{
\packageDescription{fMM}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{fMM}
\packageIndices{fMM}
Efficient matix mutliplication for big datasets.
}
\author{
\packageAuthor{fMM}
Maintainer: \packageMaintainer{fMM}
}
\references{
Find reference.
}
\keyword{ package }
\examples{
}
