\name{FMM-package}

\alias{FMM-package}
\alias{FMM}

\docType{package}

\title{Rhythmic Patterns Modeling by FMM Models}


\description{
    Provides a collection of functions to fit and explore single, multi-component and restricted Frequency Modulated Moebius (FMM) models. 'FMM' is a nonlinear parametric regression model capable of fitting non-sinusoidal shapes in rhythmic patterns.
}


\details{
  \preformatted{
     Package:    FMM
     Type:       Package
     Version:    0.1.1
     Date:       2021
     License:    GPL Version 2 or later
    }
    For a complete list of functions with individual help pages, use \code{library(help = "FMM")}.
    }

\author{
   Adrian Lamela, Itziar Fernandez, Yolanda Larriba, Alejandro Rodriguez, Cristina Rueda

   Maintainer: Itziar Fernandez <itziar.fernandez@uva.es>
}

\references{
Rueda C, Larriba Y, Peddada SD (2019).
   Frequency Modulated Moebius Model Accurately Predicts Rhythmic Signals in Biological and Physical Sciences.
   \emph{Scientific reports}, \bold{9} (1), 18701. \url{https://www.nature.com/articles/s41598-019-54569-1}
}

\keyword{ package }
